% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.pareg}
\alias{as.data.frame.pareg}
\title{as.data.frame for an object of class \code{pareg}.}
\usage{
\method{as.data.frame}{pareg}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{pareg}.}

\item{row.names}{Optional character vector of rownames.}

\item{optional}{Allow optional arguments.}

\item{...}{Additional arguments.}
}
\value{
Dataframe containing enrichment score and name for each pathway.
}
\description{
Retrieve dataframe with enrichment information.
}
\examples{
df_genes <- data.frame(
  gene = paste("g", 1:20, sep = ""),
  pvalue = c(
    rbeta(10, .1, 1),
    rbeta(10, 1, 1)
  )
)
df_terms <- rbind(
  data.frame(
    term = "foo",
    gene = paste("g", 1:10, sep = "")
  ),
  data.frame(
    term = "bar",
    gene = paste("g", 11:20, sep = "")
  )
)
fit <- pareg(df_genes, df_terms, max_iterations = 10)
as.data.frame(fit)
}
