% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_dense_sim}
\alias{as_dense_sim}
\title{Convert matrices.}
\usage{
as_dense_sim(mat_sparse)
}
\arguments{
\item{mat_sparse}{Sparse matrix.}
}
\value{
Dense matrix
}
\description{
Convert sparse similarity matrix from package data to a
dense version with 1 on its diagonal.
This matrix can then be used by \code{pareg}.
}
\examples{
transform_y(c(0, 0.5, 1))
}
