% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_enrichplot_object}
\alias{as_enrichplot_object}
\title{Convert object of class \code{pareg} to class \code{enrichResult}.}
\usage{
as_enrichplot_object(x, pvalue_threshold = 0.05)
}
\arguments{
\item{x}{An object of class \code{pareg}.}

\item{pvalue_threshold}{Treshold to select genes for count statistics.}
}
\value{
Object of class \code{enrichResult}.
}
\description{
The resulting object can be passed to any method from the
 enrichplot package and thus allows for nice visualizations of the
 enrichment results. Note: term similarities are included if available.
}
\examples{
df_genes <- data.frame(
  gene = paste("g", 1:20, sep = ""),
  pvalue = c(
    rbeta(10, .1, 1),
    rbeta(10, 1, 1)
  )
)
df_terms <- rbind(
  data.frame(
    term = "foo",
    gene = paste("g", 1:10, sep = "")
  ),
  data.frame(
    term = "bar",
    gene = paste("g", 11:20, sep = "")
  )
)
fit <- pareg(df_genes, df_terms, max_iterations = 10)
as_enrichplot_object(fit)
}
