% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_execution.R
\name{cluster_apply}
\alias{cluster_apply}
\title{Parallelize function calls on LSF cluster.}
\usage{
cluster_apply(
  df_iter,
  func,
  .bsub_params = c("-n", "2", "-W", "24:00", "-R", "rusage[mem=10000]"),
  .tempdir = ".",
  .packages = c(),
  ...
)
}
\arguments{
\item{df_iter}{Dataframe over whose rows to iterate.}

\item{func}{Function to apply to each dataframe row.
Its arguments must be all dataframe columns.}

\item{.bsub_params}{Parameters to pass to `bsub` during job submission.}

\item{.tempdir}{Location to store auxiliary files in.}

\item{.packages}{Packages to import in each job.}
}
\value{
Dataframe created by concatenating results of each function call.
}
\description{
Run function for each row of input dataframe
 in LSF job.
}
\examples{
\dontrun{
foo <- 42
cluster_apply(
  data.frame(i = 1:3, group = c('A', 'B', 'C')),
  function(i, group) {
    log_debug("hello")
    data.frame(group = group, i = i, foo = foo, result = foo + 2 * i)
  },
  .packages = c(logger)
)
}
}
