% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_similarities.R
\name{compute_term_similarities}
\alias{compute_term_similarities}
\title{Term similarity computation.}
\usage{
compute_term_similarities(
  df_terms,
  similarity_function = jaccard,
  max_similarity = 1
)
}
\arguments{
\item{df_terms}{Dataframe storing pathway database.}

\item{similarity_function}{Function to compute similarity between two sets.}

\item{max_similarity}{Value to fill diagonal with.}
}
\value{
Symmetric matrix of similarity scores.
}
\description{
Generate similarity matrix for input terms.
}
\examples{
df_terms <- data.frame(
  term = c("A", "A", "B", "B", "B", "C", "C", "C"),
  gene = c("a", "b", "a", "b", "c", "a", "c", "d")
)
compute_term_similarities(df_terms)
}
