% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_model_df}
\alias{create_model_df}
\title{Create design matrix.}
\usage{
create_model_df(df_genes, df_terms, pvalue_threshold = 0.05)
}
\arguments{
\item{df_genes}{Dataframe storing gene names and DE p-values.}

\item{df_terms}{Dataframe storing pathway database.}
}
\value{
Dataframe.
}
\description{
Store term membership for each gene.
}
\examples{
df_genes <- data.frame(
  gene = c("g1", "g2"),
  pvalue = c(0.1, 0.2)
)
df_terms <- data.frame(
  term = c("A", "A", "B", "B", "C"),
  gene = c("g1", "g2", "g1", "g2", "g2")
)
create_model_df(df_genes, df_terms)
}
