% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netreg.R
\docType{methods}
\name{family}
\alias{family}
\alias{gaussian}
\alias{bernoulli}
\alias{beta}
\alias{beta_phi_lm}
\alias{beta_phi_var}
\title{Family objects for models}
\usage{
family(object, ...)

gaussian(link = c("identity"))

bernoulli(link = c("logit", "probit", "log"))

beta(link = c("logit", "probit", "log"))

beta_phi_lm(link = c("logit", "probit", "log"))

beta_phi_var(link = c("logit", "probit", "log"))
}
\arguments{
\item{object}{a object for which the family shoulr be retured
(e.g. \code{edgenet})}

\item{...}{further arguments passed to methods}

\item{link}{name of a link function}
}
\value{
An object of class \code{pareg.family}
 \item{family }{ name of the family}
 \item{link }{ name of the link function}
 \item{linkinv }{ inverse link function}
 \item{loss }{ loss function}
}
\description{
Family objects provide a convenient way to specify the details
 of the models used by \code{pareg}.
 See also \code{\link[stats:family]{stats::family}} for more details.
}
\examples{
gaussian()
binomial("probit")$link
poisson()$linkinv
gamma()$linkinv
beta()$loss
inverse.gaussian()$loss
}
