% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_similarity_matrix}
\alias{generate_similarity_matrix}
\title{Similarity matrix generation.}
\usage{
generate_similarity_matrix(cluster_sizes)
}
\arguments{
\item{cluster_sizes}{List of cluster sizes.}
}
\value{
Similarity matrix with samples as row-/colnames.
}
\description{
Generate block-structured similarity matrices
corresponding to cluster structures.
}
\examples{
generate_similarity_matrix(c(1, 2, 3))
}
