% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_similarities.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard similarity.}
\usage{
jaccard(x, y)
}
\arguments{
\item{x}{First set.}

\item{y}{Second set.}
}
\value{
Jaccard similarity between set x and y.
}
\description{
Compute Jaccard similarity between two sets.
}
\examples{
jaccard(c(1, 2, 3), c(2, 3, 4))
}
\seealso{
Other pathway similarity methods: 
\code{\link{overlap_coefficient}()}
}
\concept{pathway similarity methods}
