% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_similarities.R
\name{overlap_coefficient}
\alias{overlap_coefficient}
\title{Overlap coefficient.}
\usage{
overlap_coefficient(x, y)
}
\arguments{
\item{x}{First set.}

\item{y}{Second set.}
}
\value{
Overlap coefficient between set x and y.
}
\description{
Compute overlap coefficient between two sets.
}
\examples{
overlap_coefficient(c(1, 2, 3), c(2, 3, 4))
}
