% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{pareg_env}
\alias{pareg_env}
\title{Conda environment definition.}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
pareg_env
}
\description{
Declare Python packages needed to run this R package.
}
\keyword{datasets}
