% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_similarities.R
\name{jaccard}
\alias{jaccard}
\alias{overlap_coefficient}
\title{Jaccard similarity.}
\usage{
jaccard(x, y)

overlap_coefficient(x, y)
}
\arguments{
\item{x}{First set.}

\item{y}{Second set.}
}
\value{
Jaccard similarity between set x and y.

Overlap coefficient between set x and y.
}
\description{
Compute Jaccard similarity between two sets.

Compute overlap coefficient between two sets.
}
\examples{
jaccard(c(1, 2, 3), c(2, 3, 4))
overlap_coefficient(c(1, 2, 3), c(2, 3, 4))
}
