% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.pareg}
\alias{plot.pareg}
\title{Plot pareg object.}
\usage{
\method{plot}{pareg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pareg}.}

\item{...}{Parameters passed to pareg::plot_pareg_with_args}
}
\value{
ggplot object.
}
\description{
Check pareg::plot_pareg_with_args for details. Needed because
of WARNING in "checking S3 generic/method consistency"
}
