% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pareg_with_args}
\alias{plot_pareg_with_args}
\title{Plot result of enrichment computation.}
\usage{
plot_pareg_with_args(x, show_term_names = TRUE, min_similarity = 0)
}
\arguments{
\item{x}{An object of class \code{pareg}.}

\item{show_term_names}{Whether to plot node labels.}

\item{min_similarity}{Don't plot edges for similarities below this value.}
}
\value{
ggplot object.
}
\description{
Visualize pathway enrichments as network.
}
\examples{
df_genes <- data.frame(
  gene = paste("g", 1:20, sep = ""),
  pvalue = c(
    rbeta(10, .1, 1),
    rbeta(10, 1, 1)
  )
)
df_terms <- rbind(
  data.frame(
    term = "foo",
    gene = paste("g", 1:10, sep = "")
  ),
  data.frame(
    term = "bar",
    gene = paste("g", 11:20, sep = "")
  )
)
fit <- pareg(df_genes, df_terms)
plot(fit)
}
