% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarity_sample}
\alias{similarity_sample}
\title{Sample elements based on similarity structure.}
\usage{
similarity_sample(sim_mat, size, similarity_factor = 1)
}
\arguments{
\item{sim_mat}{Similarity matrix with samples as row/col names.}

\item{size}{How many samples to draw.}

\item{similarity_factor}{Uniform sampling for 0. Weights mixture of uniform
and similarity vector for each draw.}
}
\value{
Vector of samples.
}
\description{
Choose similar object more often,
depending on `similarity_factor`.
}
\examples{
similarity_sample(matrix(runif(100), nrow = 10, ncol = 10), 3)
}
