% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_y}
\alias{transform_y}
\title{Transform vector from [0, 1] to (0, 1).}
\usage{
transform_y(y)
}
\arguments{
\item{y}{Numeric vector in [0, 1]^N}
}
\value{
Numeric vector in (0, 1)^N
}
\description{
Make (response) vector conform to Beta assumptions
as described in section 2 of the betareg vignette
https://cran.r-project.org/web/packages/betareg/vignettes/betareg.pdf.
}
\examples{
transform_y(c(0, 0.5, 1))
}
