% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway_similarities}
\alias{pathway_similarities}
\title{Collection of pathway similarity matrices.}
\format{
A list of lists of matrices.
* Pathway database 1
  * Similarity measure 1
  * Similarity measure 2
  * ...
* Pathway database 2
* ...
}
\usage{
pathway_similarities
}
\description{
Contains matrices for various pathway databases and similarity measures.
Note that the matrices are sparse and upper triangular.
To use with \code{pareg}, transform them using \code{pareg::as_dense_sim}.
}
\keyword{datasets}
